requirements.txt
# 核心依赖
python-dotenv>=0.19.0
requests>=2.26.0
loguru>=0.6.0

# 文本处理
nltk>=3.6.5
spacy>=3.2.0
scikit-learn>=1.0.0
pandas>=1.3.4
numpy>=1.21.4

# 大模型API
openai>=0.27.0

# 可选增强
pytz>=2021.3
tqdm>=4.62.3